refactor: change file name, time response, return type at /services/FilterManager.ts
