+++ src/api/setup.ts
@@ -5,7 +5,7 @@ import { AxiosInstance } from 'axios';
 // This function will create instance of axios for operate http request and each request will limit 10 default items
 export const instanceAxios: AxiosInstance = axios.create({
   baseURL: import.meta.env.VITE_API_URL,
-  timeout: 30000,
+  timeout: 3000,
   headers: {
     'Content-Type': 'application/json',
   },
