+++ src/helpers/checkboxEvent.ts
