+++ src/helpers/showInfo.ts
