+++ src/views/components/Form/EditUser.ts
@@ -1,8 +1,8 @@
-import { UserController } from '@/controllers/users.controller';
-import { getFormData } from '@/helpers/formData.helpers';
-import { showNotification } from '@/helpers/showNotification.helpers';
+import { UserController } from '@/controllers/users';
+import { getFormData } from '@/helpers/formData';
+import { showNotification } from '@/helpers/showNotification';
 import { BaseUserForm } from './BaseUserForm';
-import { AmountUtils } from '@/views/utils/Amount.utils';
+import { AmountUtils } from '@/views/utils/Amount';
 import { Router } from '@/routers/Router';
 
 export class EditUser extends BaseUserForm {
