+++ src/views/components/Pagination/Info.ts
@@ -1,4 +1,4 @@
-import { DOM } from '@/views/utils/DOM.utils';
+import { DOM } from '@/views/utils/DOM';
 const optionValue = ['10', '20', '50'];
 
 export class Info {
