+++ src/views/components/Table/Menu/TableOption.ts
@@ -1,4 +1,4 @@
-import { DOM } from '../../../utils/DOM.utils';
+import { DOM } from '../../../utils/DOM';
 import { ButtonCreate } from './ButtonCreate';
 import { ButtonPay } from './ButtonPay';
 import { Filter } from './Filter';
