+++ src/views/components/Table/TableData/TableData.ts
@@ -1,12 +1,12 @@
-import { UserResponse } from '@/controllers/users.controller';
+import { UserResponse } from '@/controllers/users';
 import { Heading } from './Heading';
 import icons from '@/constant/icons';
-import { DOM } from '@/views/utils/DOM.utils';
+import { DOM } from '@/views/utils/DOM';
 import { Activity } from './Activity';
-import { checkboxEvent } from '@/helpers/checkboxEvent.helpers';
-import { showInfoEvent } from '@/helpers/showInfo.helpers';
+import { checkboxEvent } from '@/helpers/checkboxEvent';
+import { showInfoEvent } from '@/helpers/showInfo';
 import { AdditionOption } from './AdditionOption';
-import { toggleHiddenClass } from '@/helpers/additionOptionEvent.helpers';
+import { toggleHiddenClass } from '@/helpers/additionOptionEvent';
 
 type StatusType = 'active' | 'payment';
 /**
